Sprint 5/Exercicios/Exercicio-TMDB/requirements.txt
requests
pandas
python-dotenv
IPython